fix: resolve all clippy warnings across all targets (0 remaining)
- Gate benchmark integration tests behind 'benchmarks' feature
- Add required-features to [[bench]] entries in Cargo.toml
- Add dhat-heap feature declaration for hotpath allocation tests
- Add #[allow] for unwrap_used/expect_used in test modules (standard practice)
- Fix module_inception in bias_feedback_tests.rs
- Fix float_cmp allows in metrics and optimization result tests
- Fix needless_update allow in async_wrapper_tests

+++ tests/benchmark_sliding_window_optimization.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "benchmarks")]
+
 use rs_vio::estimator::{Frame, SlidingWindow};
 use rs_vio::feature_tracker::Feature;
 use std::time::Instant;
@@ -8,7 +10,7 @@ fn bench_optimization_build() {
 
     // Add 8 frames
     for i in 0..8 {
-        let mut frame = Frame::new((i * 100) as i64, i);
+        let mut frame = Frame::new((i * 100) as i64, i as i32);
         frame.is_keyframe = true;
 
         // Add 2000 features per frame to stress test
